Fixing RbGame
Also migrating roblox commands to new DSL
