+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/commands/CommandManager.kt
@@ -12,8 +12,6 @@ import com.mrpowergamerbr.loritta.commands.vanilla.minecraft.*
 import com.mrpowergamerbr.loritta.commands.vanilla.misc.*
 import com.mrpowergamerbr.loritta.commands.vanilla.music.LyricsCommand
 import com.mrpowergamerbr.loritta.commands.vanilla.pokemon.PokedexCommand
-import com.mrpowergamerbr.loritta.commands.vanilla.roblox.RbGameCommand
-import com.mrpowergamerbr.loritta.commands.vanilla.roblox.RbUserCommand
 import com.mrpowergamerbr.loritta.commands.vanilla.social.*
 import com.mrpowergamerbr.loritta.commands.vanilla.undertale.UndertaleBattleCommand
 import com.mrpowergamerbr.loritta.commands.vanilla.undertale.UndertaleBoxCommand
@@ -169,10 +167,6 @@ class CommandManager(loritta: Loritta) {
 		commandMap.add(McConquistaCommand())
 		commandMap.add(McSkinCommand())
 		commandMap.add(McMoletomCommand())
-
-		// =======[ ROBLOX ]========
-		commandMap.add(RbUserCommand())
-		commandMap.add(RbGameCommand())
 
 		// =======[ UNDERTALE ]========
 		commandMap.add(UndertaleBoxCommand())
